Split out cell components into own files
